+++ src/main.js
@@ -62,7 +62,7 @@ class PlatformIOVSCodeExtension {
         outputChannel.show();
 
         outputChannel.appendLine('Installing PlatformIO Core...');
-        outputChannel.appendLine("Please don't close this window and don't "
+        outputChannel.appendLine('Please don\'t close this window and don\'t '
           + 'open other folders until this process is completed.');
 
         try {
@@ -90,22 +90,37 @@ class PlatformIOVSCodeExtension {
     ));
     this._context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.build',
-      () => vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Build')
+      async () => {
+        await this.terminateUploadTask();
+        vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Build');
+      }
     ));
     this._context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.upload',
-      () => {
-        this.pioTerm.closeSerialMonitor();
-        vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Upload');
+      async () => {
+        await this.terminateUploadTask();
+
+        let task = 'PlatformIO: Upload';
+        const config = vscode.workspace.getConfiguration('platformio-ide');
+        if(config.get('forceUploadAndMonitor')) {
+          task = 'PlatformIO: Upload and Monitor';
+        }
+        vscode.commands.executeCommand('workbench.action.tasks.runTask', task);
       }
     ));
     this._context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.clean',
-      () => vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Clean')
+      async () => {
+        await this.terminateUploadTask();
+        vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Clean');
+      }
     ));
     this._context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.serialMonitor',
-      () => this.pioTerm.sendText('pio device monitor')
+      async () => {
+        await this.terminateUploadTask();
+        vscode.commands.executeCommand('workbench.action.tasks.runTask', 'PlatformIO: Monitor');
+      }
     ));
     this._context.subscriptions.push(vscode.commands.registerCommand(
       'platformio-ide.libraryManager',
@@ -115,6 +130,15 @@ class PlatformIOVSCodeExtension {
       'platformio-ide.newTerminal',
       () => this.pioTerm.new().show()
     ));
+  }
+
+  async terminateUploadTask() {
+    try {
+      await vscode.commands.executeCommand('workbench.action.tasks.terminate');
+    } catch(err) {
+      console.error(err);
+    }
+    return new Promise(resolve => setTimeout(() => resolve(), 500));
   }
 
   initProjectIndexer() {